scripts/requirements.txt

requests
tqdm
pillow
